exoplayer: upd sabr classes

+++ exoplayer-amzn-2.10.6/library/sabr/src/main/java/com/google/android/exoplayer2/source/sabr/parser/exceptions/MediaSegmentMismatchError.java
@@ -3,10 +3,10 @@ package com.google.android.exoplayer2.source.sabr.parser.exceptions;
 import com.google.android.exoplayer2.source.sabr.protos.videostreaming.FormatId;
 
 public class MediaSegmentMismatchError extends SabrStreamError {
-    public final int expectedSequenceNumber;
-    public final int receivedSequenceNumber;
+    public final long expectedSequenceNumber;
+    public final long receivedSequenceNumber;
 
-    public MediaSegmentMismatchError(FormatId formatId, int expectedSequenceNumber, int receivedSequenceNumber) {
+    public MediaSegmentMismatchError(FormatId formatId, long expectedSequenceNumber, long receivedSequenceNumber) {
         super(String.format(
                 "Segment sequence number mismatch for format %s: expected %s, received %s",
                 formatId,
